fixes from rename
